fix(workflow): Ensure artifacts directory exists before creating tarballs
This commit modifies the "Create Tarball" step in `apt.yml` to ensure the `artifacts` directory is created at the beginning of the step, before any tarball creation commands are executed.

Previously, the `artifacts` directory was created later in the "Create Tarball" step, after the attempt to create `siphon_${{ env.VERSION }}.orig.tar.gz`. This caused the `tar` command to fail with "Cannot open: No such file or directory" because the target `artifacts` directory did not yet exist when the command was executed.

This commit moves the `mkdir artifacts` command to the very beginning of the "Create Tarball" step, guaranteeing that the `artifacts` directory is available when the `tar` command attempts to write the `siphon_${{ env.VERSION }}.orig.tar.gz` file, resolving the file creation error.

Refs yuezk/publish-ppa-package#8