fix(apt-workflow): correct PPA publish tarball path to match created artifact
The PPA publish step was failing because it was attempting to locate
'siphon_${{ env.VERSION }}.tar.gz', which does not exist. The tarball
artifact created is 'siphon_${{ env.VERSION }}.orig.tar.gz'. This commit
corrects the path in the PPA publish step to match the created tarball,
ensuring successful publication.