ci: Add error handling to tarball upload step
This commit adds error handling to the tarball upload step in the `apt.yml` workflow.

The `if-no-files-found: error` option is now used to ensure that the workflow stops with an error if the tarball file (`siphon_${{ env.VERSION }}.orig.tar.gz`) is not found. This prevents the workflow from proceeding with an invalid or missing artifact.