fix(apt-workflow): Revert to yuezk/publish-ppa-package@main for PPA publishing
Revert action version to `main` branch to address potential issues with `v2`.

This commit changes the "Publish to PPA" step in the APT workflow back to using the `yuezk/publish-ppa-package@main` action instead of `yuezk/publish-ppa-package@v2`.

This rollback is intended to mitigate potential issues encountered with the `v2` version of the action, specifically the "cd: too many arguments" error observed in recent workflow runs.  Using the `main` branch may provide a more stable or compatible version of the PPA publishing action.